feat: extract toolbar
