+++ src/frontend/src/eth/components/tokens/AddTokenForm.svelte
@@ -1,16 +1,13 @@
 <script lang="ts">
 	import { Input } from '@dfinity/gix-components';
-	import { createEventDispatcher } from 'svelte';
 	import { isNullishOrEmpty } from '$lib/utils/input.utils';
 	import { i18n } from '$lib/stores/i18n.store';
-	import ButtonGroup from '$lib/components/ui/ButtonGroup.svelte';
+	import AddTokenByNetworkToolbar from '$icp-eth/components/tokens/AddTokenByNetworkToolbar.svelte';
 
 	export let contractAddress = '';
 
 	let invalid = true;
 	$: invalid = isNullishOrEmpty(contractAddress);
-
-	const dispatch = createEventDispatcher();
 </script>
 
 <div class="stretch pt-2">
@@ -25,11 +22,4 @@
 	/>
 </div>
 
-<ButtonGroup>
-	<button type="button" class="secondary block flex-1" on:click={() => dispatch('icBack')}
-		>{$i18n.core.text.back}</button
-	>
-	<button class="primary block flex-1" type="submit" disabled={invalid} class:opacity-10={invalid}>
-		{$i18n.core.text.next}
-	</button>
-</ButtonGroup>
+<AddTokenByNetworkToolbar {invalid} on:icBack />
